requirements.txt
absl-py==1.4.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.5
cycler==0.11.0
dj-database-url==2.0.0
Django==4.0.6
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.37.2
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.0
ml-dtypes==0.1.0
mpmath==1.2.1
networkx==3.0
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sqlparse==0.4.2
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
wrapt==1.14.1
zipp==3.15.0
